sstable: split Writer into Writer and RawWriter
The sstable.Writer had some egronomic methods for writing to the Writer from
non-internal client code (implementing the pebble.Writer interface). These
methods are internally unused, but may be used by external clients writing to
sstables for ingestion, storage outside the LSM, tests, etc. This commit splits
the Writer into two types, with the sugaring methods on the Writer type and the
lower-level internal methods on the RawWriter type.

In future work, sstable.RawWriter may become an interface to support two
implementations: one for rowblk-based sstables and one for colblk-based
sstables.
